Fix HasHighlyAvailableControlPlane to use AllInstanceGroups
When using 'kops update cluster' with --instance-group or
--instance-group-roles filters, the HasHighlyAvailableControlPlane
function was incorrectly using the filtered InstanceGroups list
instead of AllInstanceGroups.

This caused cluster-wide controllers like aws-load-balancer-controller
and node-termination-handler to incorrectly downscale replicas from 2
to 1 when updating a specific instance group in an HA cluster.

The fix changes HasHighlyAvailableControlPlane to use AllInstanceGroups
since HA status is a cluster-wide property, not specific to filtered
instance groups.

Signed-off-by: Peter Kubicsek <peter.kubicsek@happening.xyz>